Moved addEscapeListener to dom utilities Fixed doc blocks
