+++ library/src/scripts/dom.ts
@@ -569,3 +569,28 @@ export function getElementHeight(
         bottomMargin: bottomHeight,
     };
 }
+
+/**
+ * Register an keyboard listener for the escape key.
+ *
+ * @param root - The element to watch for the escape listener in.
+ * @param returnElement - The element to return to when escape is pressed.
+ * @param callback
+ */
+export function addEscapeListener(
+    root: HTMLElement,
+    returnElement: HTMLElement,
+    callback: (event: KeyboardEvent) => void = () => {
+        return;
+    },
+) {
+    root.addEventListener("keydown", (event: KeyboardEvent) => {
+        if (event.key === "Escape") {
+            if (root.contains(document.activeElement)) {
+                event.preventDefault();
+                returnElement.focus();
+                callback(event);
+            }
+        }
+    });
+}
